ux: (perf) use ws for frontend-backend comms
Attempted to use rn_bridge.channel, but figured that it has UDP-like
properties (quite fast but doesn't guarantee order and reliability), so
used websockets instead, which RN supports for the frontend thread to
talk to the backend thread. It uses multiserver noauth transform plugin
to avoid crypto shs between frontend and backend.