dx: [ios] disable Rust node modules in the backend
ssb-keys-mnemonic-neon had some transitive dependencies updated so
that it doesn't compile anymore on old versions of Rust (e.g. 1.46.0),
so this means we are forced to use new versions of Rust, but in new
versions we have issue https://github.com/rust-lang/rust/issues/21727
which means that Rust and Cargo do not support cdylib, which is really
what we need on iOS. So, we're stuck, and this commit disables Rust
usage on iOS just so we can make working releases for iOS users.