ux: [and] reduce APK size and support Android 5.0
Closes issue #1400 and #1499 simultaneously.

The cause for the large APK size was strange: if you set the minimum
supported Android version to Android 6.0, then the APK size jumps to
100MB. If you set the minimum to Android 5.0, the size goes back to
48MB.

Related issue: https://github.com/facebook/react-native/issues/30284

This means we have an easy solution, which is just to drop the minSdk
to 21, but that will mean that issue #1400 would come back (because
that crash is specific to Android 5.0 i.e. 21). One way forward would
be to disable Rust for Android too, then both #1499 and #1400 is fixed.

Unfortunately this means we don't have Rust running anymore, but it's
higher priority to fix showstopper crashes and APK sizes than it is
to have Rust's marginal performance improvement.